Addition of Domain and Services Projects
