+++ LibraryApp/LibraryApp.Api/Controllers/LibraryController.cs
@@ -1,4 +1,5 @@
-﻿using LibraryApp.Console.Services;
+﻿using LibraryApp.Domain;
+using LibraryApp.Services;
 using Microsoft.AspNetCore.Mvc;
 
 namespace LibraryApp.Api.Controllers
@@ -30,11 +31,24 @@ namespace LibraryApp.Api.Controllers
         }
 
         // Add POST to add a new book
-        [HttpPost("book")]
-        public IActionResult PostBook(string title, string author, int pages = 0)
+        //[HttpPost("book")]
+        //// My Version
+        //public IActionResult PostBook(string title, string author, int pages = 0)
+        //{
+        //    var book = _service.AddBook(title, author, pages);
+        //    return Ok(book);
+        //}
+
+        // Mike's Version
+        [HttpPost("books")]
+        public IActionResult AddBook([FromBody] Book book)
         {
-            var book = _service.AddBook(title, author, pages);
-            return Ok(book);
+            if (book == null || string.IsNullOrEmpty(book.Title) || string.IsNullOrEmpty(book.Author))
+            {
+                return BadRequest("Invalid book data.");
+            }
+            var addedBook = _service.AddBook(book.Title, book.Author, book.Pages);
+            return CreatedAtAction(nameof(GetItems), new { id = addedBook.Id }, addedBook);
         }
     }
 }
