+++ LibraryApp/LibraryApp.Api/Program.cs
@@ -1,9 +1,4 @@
-using Microsoft.AspNetCore.Builder;
-using Swashbuckle.AspNetCore.SwaggerGen;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.OpenApi.Models;
-using Swashbuckle.AspNetCore.SwaggerUI;
-using LibraryApp.Console.Services;
+using LibraryApp.Services;
 
 var builder = WebApplication.CreateBuilder(args);
 
@@ -12,7 +7,7 @@ var builder = WebApplication.CreateBuilder(args);
 builder.Services.AddControllers();
 // Learn more about configuring OpenAPI at https://aka.ms/aspnet/openapi
 builder.Services.AddOpenApi();
-builder.Services.AddEndpointsApiExplorer();
+builder.Services.AddEndpointsApiExplorer(); 
 builder.Services.AddSwaggerGen();
 // Add CORS services
 builder.Services.AddCors(options =>
