+++ LibraryApp/LibraryApp.Console/Program.cs
@@ -1,6 +1,7 @@
-﻿using LibraryApp.Console.Domain;
-using LibraryApp.Console.Services;
-using LibraryApp.Console.Utils;
+﻿using LibraryApp.Console.Utils;
+using LibraryApp.Domain;
+using LibraryApp.Services;
+
 public class Program
 {
     private static readonly LibraryService _service = new();
