+++ LibraryApp/LibraryApp.Domain/Book.cs
@@ -4,22 +4,22 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace LibraryApp.Console.Domain
+namespace LibraryApp.Domain
 {
     public class Book : LibraryItem
     {
         public string Author { get; }
         public int Pages { get; }
         // Constructor chaining example:
-        public Book(int id, string title, string author) : this(id, title, author, pages: 0) { } 
+        public Book(int id, string title, string author) : this(id, title, author, pages: 0) { }
         public Book(int id, string title, string author, int pages) : base(id, title)
         {
             Author = string.IsNullOrEmpty(author) ? "unknown" : author.Trim();
             Pages = pages < 0 ? 0 : pages;
         }
 
-        public override string GetInfo() 
+        public override string GetInfo()
             => $"Book [Id={Id}, Title={Title}, Author={Author} Pages={Pages}, IsBorrowed={IsBorrowed}]";
-        
+
     }
 }
