+++ LibraryApp/LibraryApp.Domain/LibraryItem.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace LibraryApp.Console.Domain
+namespace LibraryApp.Domain
 {
     public abstract class LibraryItem
     {
