+++ LibraryApp/LibraryApp.Services/LibraryApp.Services.csproj
@@ -1,7 +1,6 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <OutputType>Exe</OutputType>
     <TargetFramework>net9.0</TargetFramework>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
@@ -9,7 +8,6 @@
 
   <ItemGroup>
     <ProjectReference Include="..\LibraryApp.Domain\LibraryApp.Domain.csproj" />
-    <ProjectReference Include="..\LibraryApp.Services\LibraryApp.Services.csproj" />
   </ItemGroup>
 
 </Project>
