+++ LibraryApp/LibraryApp.Services/LibraryService.cs
@@ -1,12 +1,11 @@
-﻿using LibraryApp.Console.Domain;
-using LibraryApp.Console.Utils;
+﻿using LibraryApp.Domain;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace LibraryApp.Console.Services
+namespace LibraryApp.Services
 {
     public sealed class LibraryService : ILibraryService
     {
