+++ LibraryApp/LibraryApp.sln
@@ -1,11 +1,15 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
-VisualStudioVersion = 17.14.36414.22 d17.14
+VisualStudioVersion = 17.14.36414.22
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LibraryApp.Console", "LibraryApp.Console\LibraryApp.Console.csproj", "{799A6395-FA2A-43CD-852C-7AA9A1957984}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LibraryApp.Api", "LibraryApp.Api\LibraryApp.Api.csproj", "{5A74E60D-C8E6-442D-B345-BAD48785C359}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LibraryApp.Services", "LibraryApp.Services\LibraryApp.Services.csproj", "{B5C64577-4535-4989-8BD7-ACB9804B466E}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LibraryApp.Domain", "LibraryApp.Domain\LibraryApp.Domain.csproj", "{9D561078-F00A-4F8B-A846-F0C76CE0D71C}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -21,6 +25,14 @@ Global
 		{5A74E60D-C8E6-442D-B345-BAD48785C359}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{5A74E60D-C8E6-442D-B345-BAD48785C359}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{5A74E60D-C8E6-442D-B345-BAD48785C359}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B5C64577-4535-4989-8BD7-ACB9804B466E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B5C64577-4535-4989-8BD7-ACB9804B466E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B5C64577-4535-4989-8BD7-ACB9804B466E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B5C64577-4535-4989-8BD7-ACB9804B466E}.Release|Any CPU.Build.0 = Release|Any CPU
+		{9D561078-F00A-4F8B-A846-F0C76CE0D71C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{9D561078-F00A-4F8B-A846-F0C76CE0D71C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{9D561078-F00A-4F8B-A846-F0C76CE0D71C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{9D561078-F00A-4F8B-A846-F0C76CE0D71C}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
